Rework main process to enable testing
This change reworks the main hause to use dependency injection so
that the client interfaces can be mocked. I added some very basic
testing to prove that the basic process is working correctly.

I plan on following up with some more structural changes to make
the configuration behavior a little more consistent.