Convert hauser main process to service and add testing (#86)
This change reworks the main hauser process to use dependency injection so
that the client interfaces can be mocked. I added some very basic
testing to prove that the basic process is working correctly.

I plan on following up with some more structural changes to make
the configuration behavior a little more consistent.