Removed warehouse dependency on ExportMeta
To prep for using the new APIs, we need to remove the dependency
`ExportMeta` from the warehouse package and instead rely directly
on the "stop time" (i.e. the time for the last bundle we laoded into
the warehouse). This change should be fairly transparent with the
exception of the ID that was inserted into the sync table.