Use new APIs for exports (#95)
This is significant update which migrates the existing export client from the existing APIs to the new (as of yet, undocumented) "Segment Export" APIs.

The aim is to be as backwards compatible as possible, but there are few notable changes from the existing workflow:

### Exports are not "bundled"
With the existing APIs, the exports were "bundled" into a time period that was defined in the "Data Export" settings for in the FullStory application. The new APIs do not have this restriction. Instead, the new APIs allows a specifying a time period for which to export data. However, the change here means that we no longer support "listing" exports since they are now user defined. Instead, a `StartTime` must be specified in the config if starting fresh. However, if a "sync" exists for the provided configuration, `StartTime` will simply be ignored.

The second implication of this is that the export duration must be specified in the config as well. The default value (if unspecified) is now 1 hour.

### New fields
With the new APIs a few new fields have been added:
* UserCreated: 
* EventTargetSelector
* SessionStart
* PageStart
* PagePlatform
* PageScreenWidth
* PageScreenHeight
* PageViewportWidth
* PageViewportHeight

### Special cased fields
* EventTargetSelectorTok: this will only be exported if a database already exists with this column.
* PageUserAgent: this field used to be called `PageAgent`. If a database already has this column, `PageUserAgent` will be translated to match. Otherwise, the column will be named `PageUserAgent`.