Fix folder permissions in docker (#107)
There was an issue with using the scratch
image because hauser expects to be able to create
directories for temporary files. With the scratch
image, the `hauser` user did not have permission to
create the directory, so the process would fail.
Since "chown" doesn't exist in the scratch image,
just use an alpine image.